+++ eoullim-Project/src/main/java/org/example/eoullimback/_common/enums/RoleType.java
@@ -1,4 +1,4 @@
-package org.example.eoullimback.user.enums;
+package org.example.eoullimback._common.enums;
 
 public enum RoleType {
     USER, MANAGER, ADMIN
