+++ eoullim-Project/src/main/java/org/example/eoullimback/qaa/Qaa.java
@@ -5,7 +5,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.example.eoullimback._common.base.BaseTimeEntity;
-import org.example.eoullimback.user.User;
+import org.example.eoullimback.user_auth.user.User;
 
 @Entity
 @Table(
