+++ eoullim-Project/src/main/java/org/example/eoullimback/user_auth/auth/AuthController.java
@@ -1,0 +1,4 @@
+package org.example.eoullimback.user_auth.auth;
+
+public class AuthController {
+}
