+++ eoullim-Project/src/main/java/org/example/eoullimback/user_auth/auth/AuthService.java
@@ -1,0 +1,4 @@
+package org.example.eoullimback.user_auth.auth;
+
+public class AuthService {
+}
