+++ eoullim-Project/src/main/java/org/example/eoullimback/user_auth/user/User.java
@@ -1,9 +1,9 @@
-package org.example.eoullimback.user;
+package org.example.eoullimback.user_auth.user;
 
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.example.eoullimback.user.enums.RoleType;
+import org.example.eoullimback._common.enums.RoleType;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
