+++ eoullim-Project/src/main/java/org/example/eoullimback/user_auth/user/UserController.java
@@ -1,0 +1,4 @@
+package org.example.eoullimback.user_auth.user;
+
+public class UserController {
+}
