+++ eoullim-Project/src/main/java/org/example/eoullimback/user_auth/user/UserRepository.java
@@ -1,0 +1,4 @@
+package org.example.eoullimback.user_auth.user;
+
+public class UserRepository {
+}
